implementation/requirements-advanced.txt
# Core RAG Agent Dependencies
pydantic-ai>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.5
python-dotenv>=1.0.0

# LLM and Embeddings
openai>=1.54.0

# Document Processing (Docling)
docling>=2.11.1
transformers>=4.40.0
torch>=2.0.0

# Re-ranking
sentence-transformers>=2.2.2

# For BM25 search
rank-bm25>=0.2.2
numpy>=1.21.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.3.0
